examples/custom-gui/streamlit/requirements.txt
altair==5.4.1
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.1.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.22.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prompt-toolkit==3.0.49
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
